Initialise Validated<T> with string for output on a form. Convert an input string to another type with Validator.Value()

+++ tests/Mundane.Tests/Tests_Validator/When_The_Async_Validation_Succeeds.cs
@@ -32,7 +32,7 @@ namespace Mundane.Tests.Tests_Validator
 
 			var result = await Validator.Validate(validator => ValueTask.FromResult(new TestModel(validator, number)));
 
-			Assert.Equal(number, (int)result.Model.Number);
+			Assert.Equal(number, result.Model.Number);
 		}
 
 		private sealed class TestModel
